backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
python-dotenv==1.0.1
PyMuPDF==1.25.3
beautifulsoup4==4.12.3
requests==2.32.3
google-generativeai==0.8.4
aiofiles==24.1.0
pydantic==2.10.4
httpx==0.28.1
